requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
langchain-huggingface
docling
python-dotenv
loguru
sentence-transformers
Pillow